moved all folder into one and organized it
