+++ WBSchool-chat/src/app/profile-page/components/account-settings/delete-modal/delete-modal.component.scss
