+++ WBSchool-chat/src/app/profile-page/components/account-settings/delete-modal/delete-modal.component.ts
@@ -4,9 +4,9 @@ import { MatDialogRef } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { catchError, throwError } from 'rxjs';
 import { AuthService } from 'src/app/auth/services/auth.service';
-import { IProfileData, IServerResponse } from 'src/app/profile-settings/interfaces/interface';
-import { IUserDeleteData } from '../interface/account-settings';
-import { AccountSettingsService } from '../service/account-settings.service';
+import { IProfileData, IServerResponse } from 'src/app/profile-page/interfaces/profile-settings';
+import { IUserDeleteData } from '../../../interfaces/account-settings';
+import { AccountSettingsService } from '../../../services/account-settings.service';
 
 @Component({
   selector: 'app-delete-modal',
