+++ WBSchool-chat/src/app/profile-page/components/profile-settings/modal-help/modal-help.component.ts
