+++ WBSchool-chat/src/app/profile-page/components/profile-settings/profile-settings.component.html
@@ -32,11 +32,16 @@
 				</ng-template> -->
 			</div>
 				<div *ngIf="toggle && selectedItem.id == i+1" class="profile-settings__menu_item_form-field">
-					<mat-form-field appearance="fill">
-						<mat-label>{{selectedItem.title}}</mat-label>
-					    <input matInput color='warn' type="text" id={{item.id}} value={{item.description}} #input>
-						<!-- <textarea name="" id="" cols="30" rows="10" value={{item.description}}></textarea> -->
-			        </mat-form-field>
+					<!-- <div *ngIf="item.id != 3; else photoInput"> -->
+						<mat-form-field appearance="fill">
+							<mat-label>{{selectedItem.title}}</mat-label>
+							<input matInput color='warn' type="text" id={{item.id}} value={{item.description}} #input>
+							<!-- <textarea name="" id="" cols="30" rows="10" value={{item.description}}></textarea> -->
+						</mat-form-field>
+					<!-- </div>
+					<ng-template #photoInput>
+
+					</ng-template> -->
 					<div class="form-field__btn-wrapper">
 						<button mat-raised-button color="primary" (click)="[addToFormData(input), toggle = !toggle]">
 							<mat-icon>done</mat-icon>
