+++ WBSchool-chat/src/app/profile-page/services/account-settings.service.ts
@@ -3,12 +3,12 @@ import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { catchError, Observable, throwError } from 'rxjs';
 import { AuthService } from 'src/app/auth/services/auth.service';
-import { IServerResponse } from 'src/app/profile-settings/interfaces/interface';
+import { IServerResponse } from 'src/app/profile-page/interfaces/profile-settings';
 import {
   IPasswordEditData,
   IPasswordOnly,
   IUserDeleteData,
-} from '../interface/account-settings';
+} from '../interfaces/account-settings';
 
 @Injectable({
   providedIn: 'root',
