+++ WBSchool-chat/src/app/profile-page/services/profile-settings.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { catchError, Observable, throwError } from 'rxjs';
-import { IProfileData, IServerResponse } from '../interfaces/interface';
+import { IProfileData, IServerResponse } from '../interfaces/profile-settings';
 
 @Injectable({
   providedIn: 'root',
@@ -27,3 +27,4 @@ export class ProfileSettingsService {
     );
   }
 }
+
